#151 - refactoring: added pckages
