+++ src/main/java/org/opengis/cite/ogcapifeatures10/conformance/crs/discovery/collection/DiscoveryCollectionCrsUri.java
@@ -1,4 +1,4 @@
-package org.opengis.cite.ogcapifeatures10.conformance.crs.discovery;
+package org.opengis.cite.ogcapifeatures10.conformance.crs.discovery.collection;
 
 import static org.opengis.cite.ogcapifeatures10.EtsAssert.assertValidCrsIdentifier;
 
@@ -27,7 +27,7 @@ import io.restassured.path.json.JsonPath;
  *
  * @author <a href="mailto:goltz@lat-lon.de">Lyn Goltz </a>
  */
-public class DiscoveryCollection extends AbstractDiscoveryCollection {
+public class DiscoveryCollectionCrsUri extends AbstractDiscoveryCollection {
 
     /**
      * Test: crs property in the collection objects in the path /collections
