+++ src/main/resources/org/opengis/cite/ogcapifeatures10/testng.xml
@@ -26,7 +26,8 @@
   <test name="CRS - Part 2">
     <packages>
       <package name="org.opengis.cite.ogcapifeatures10.conformance.crs" />
-      <package name="org.opengis.cite.ogcapifeatures10.conformance.crs.discovery" />
+      <package name="org.opengis.cite.ogcapifeatures10.conformance.crs.discovery.collection" />
+      <package name="org.opengis.cite.ogcapifeatures10.conformance.crs.discovery.collections" />
       <package name="org.opengis.cite.ogcapifeatures10.conformance.crs.query.crs.features" />
       <package name="org.opengis.cite.ogcapifeatures10.conformance.crs.query.crs.feature" />
       <package name="org.opengis.cite.ogcapifeatures10.conformance.crs.query.bboxcrs" />
